+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.20.0...master)
+### Added
+- Added a checkbox to enable/disable median line on hydrograph 
+
 ### Changed
 - Conditionally displayed the method description based on the number of methods.
 - Updated to use wdfn-viz which updates the USWDS version to 2.2.1